requirements.txt
xai-sdk[async]>=1.3.1
streamlit>=1.38.0
python-dotenv>=1.0.1
requests>=2.31.0
Pillow>=10.0.0
graphviz>=0.20.3
ruff>=0.6.0
black>=24.0.0
mypy>=1.11.0